LayoutTests/http/wpt/wk-web-animations/timing-model/animation-effect-unique-relationship.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check an AnimationEffect can only be used for one Animation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
    const effectA = new KeyframeEffect(document.createElement("div"));
    const effectB = new KeyframeEffect(document.createElement("div"));

    const a = new Animation(effectA);
    const b = new Animation(effectB);

    assert_equals(a.effect, effectA);
    assert_equals(b.effect, effectB);

    b.effect = effectA;
    assert_equals(a.effect, null);
    assert_equals(b.effect, effectA);
}, "Check that setting an animation's effect onto another animation sets effect to null.");

</script>
</body>
